+++ ruma-events/src/lib.rs
@@ -149,6 +149,7 @@ pub mod forwarded_room_key;
 pub mod fully_read;
 pub mod ignored_user_list;
 pub mod key;
+pub mod pdu;
 pub mod presence;
 pub mod push_rules;
 pub mod receipt;