+++ ruma-federation-api/src/lib.rs
@@ -7,6 +7,5 @@ mod serde;
 pub mod directory;
 pub mod discovery;
 pub mod membership;
-pub mod pdu;
 pub mod query;
 pub mod transactions;